requirements.txt
pandas
psycopg2-binary
requests
pytz
numpy
apscheduler